// src/typechain/contracts/fwd/EssentialForwarder.ts
export declare namespace IForwardRequest {
  export type ERC721ForwardRequestStruct = {
    from: PromiseOrValue<string>;
    authorizer: PromiseOrValue<string>;
    to: PromiseOrValue<string>;
    nftContract: PromiseOrValue<string>;
    nftTokenId: PromiseOrValue<BigNumberish>;
    nftChainId: PromiseOrValue<BigNumberish>;
    targetChainId: PromiseOrValue<BigNumberish>;
    value: PromiseOrValue<BigNumberish>;
    gas: PromiseOrValue<BigNumberish>;
    nonce: PromiseOrValue<BigNumberish>;
    data: PromiseOrValue<BytesLike>;
  };

  export type ERC721ForwardRequestStructOutput = [
    string,
    string,
    string,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    string
  ] & {
    from: string;
    authorizer: string;
    to: string;
    nftContract: string;
    nftTokenId: BigNumber;
    nftChainId: BigNumber;
    targetChainId: BigNumber;
    value: BigNumber;
    gas: BigNumber;
    nonce: BigNumber;
    data: string;
  };

  export type PlaySessionStruct = {
    authorized: PromiseOrValue<string>;
    expiresAt: PromiseOrValue<BigNumberish>;
  };

  export type PlaySessionStructOutput = [string, BigNumber] & {
    authorized: string;
    expiresAt: BigNumber;
  };
}